Handle aborted rates requests gracefully
